requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==2.2.5
djangorestframework==3.10.3
idna==2.8
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
gunicorn==19.9.0
